requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pandas==2.2.2
weasyprint==62.3
python-dotenv==1.0.1
jinja2==3.1.4